combined navbar files
