+++ components/layout.tsx
@@ -1,14 +1,10 @@
 import { useTranslation } from "next-i18next"
 import Head from "next/head"
 import React, { FC, useEffect, useState } from "react"
-import { useMediaQuery } from "usehooks-ts"
 import { signOutAndRedirectToHome, useAuth } from "./auth"
 import AuthModal from "./auth/AuthModal"
-import { DesktopNav } from "./DesktopNav"
 import PageFooter from "./Footer/Footer"
-import { MobileNav } from "./MobileNav"
-
-import { useContext } from "react"
+import { MainNavbar } from "./Navbar"
 import { TabContext, TabStatus } from "./shared/ProfileTabsContext"
 
 export const PageContainer: FC<React.PropsWithChildren<unknown>> = ({
@@ -26,7 +22,6 @@ export const Layout: React.FC<React.PropsWithChildren<LayoutProps>> = ({
   title
 }) => {
   const { authenticated, user } = useAuth()
-  const isMobile = useMediaQuery("(max-width: 768px)")
   const { t } = useTranslation("common")
   const formattedTitle = title
     ? `${title} | ${t("maple_abbr")}: ${t("maple_fullName")}`
@@ -51,7 +46,7 @@ export const Layout: React.FC<React.PropsWithChildren<LayoutProps>> = ({
           </Head>
           <TabContext.Provider value={{ tabStatus, setTabStatus }}>
             <PageContainer>
-              {isMobile ? <MobileNav /> : <DesktopNav />}
+              <MainNavbar />
               <AuthModal />
               <div className={`col`}>{children}</div>
               <PageFooter
